requirements.txt
pydot
Django
binwalk
r2pipe
python-magic
django-cors-headers
